BICAS: Update test code to matlab.unittest
modified:   +bicas/+proc/dwns.m              # Test-related minor bugfix.
deleted:    +bicas/+proc/dwns___ATEST_.m
new file:   +bicas/+proc/dwns___UTEST.m
deleted:    +bicas/+utils/get_bin_indices___ATEST_.m
new file:   +bicas/+utils/get_bin_indices___UTEST.m
deleted:    +bicas/+utils/modif_std_deviation___ATEST_.m
new file:   +bicas/+utils/modif_std_deviation___UTEST.m
deleted:    +bicas/+utils/parse_CLI_options___ATEST_.m
new file:   +bicas/+utils/parse_CLI_options___UTEST.m
deleted:    +bicas/interpret_config_file___ATEST_.m
new file:   +bicas/interpret_config_file___UTEST.m

+++ mission/solar_orbiter/bicas/src/+bicas/+proc/dwns.m
@@ -285,7 +285,7 @@ classdef dwns
                 
                 zvBinsTt2000    = int64(ones(0,1));
                 iRecordsInBinCa = cell(0,1);
-                binSizeArrayNs  = zeros(0,1);
+                binSizeArrayNs  = int64(zeros(0,1));
                 return
             end
             % CASE: zvAllTt2000 is not empty.
